+++ htindex/cmd/root.go
@@ -1,23 +1,3 @@
-// Copyright © 2019 Dmitry Mozzherin <dmozzherin@gmail.com>
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-
 package cmd
 
 import (
@@ -39,10 +19,11 @@ var (
 )
 
 type config struct {
-	Root   string
-	Input  string
-	Output string
-	Jobs   int
+	Root      string
+	Input     string
+	Output    string
+	Jobs      int
+	ReportNum int
 }
 
 // rootCmd represents the base command when called without any subcommands
@@ -52,14 +33,15 @@ var rootCmd = &cobra.Command{
 	Long: `Hathi Trust is a large collection of public and private textual
 	information. The htindex program allows to use its data to find in it
 	scientific names.`,
-	// Uncomment the following line if your bare application
-	// has an action associated with it:
 	Run: func(cmd *cobra.Command, args []string) {
 		versionFlag(cmd)
 		opts = getOpts()
 		opts = getFlags(opts, cmd)
-		hti := htindex.NewHTindex(opts...)
-		err := hti.Run()
+		hti, err := htindex.NewHTindex(opts...)
+		if err != nil {
+			log.Fatal(err)
+		}
+		err = hti.Run()
 		if err != nil {
 			log.Fatal(err)
 		}
@@ -145,6 +127,12 @@ func getOpts() []htindex.Option {
 	}
 	if cfg.Jobs > 0 {
 		opts = append(opts, htindex.OptJobs(cfg.Jobs))
+	}
+	if cfg.Jobs > 0 {
+		opts = append(opts, htindex.OptJobs(cfg.Jobs))
+	}
+	if cfg.ReportNum > 0 {
+		opts = append(opts, htindex.OptReportNum(cfg.ReportNum))
 	}
 	return opts
 }