+++ runner.go
@@ -8,15 +8,15 @@ import (
 )
 
 func (hti *HTindex) Run() error {
-	fmt.Println(hti)
+	fmt.Printf("Processing with %d 'threads'\n", hti.jobsNum)
 	inCh := make(chan string)
 	errCh := make(chan error)
 	outCh := make(chan *title)
 	var wg sync.WaitGroup
 	var wgOut sync.WaitGroup
 	wg.Add(hti.jobsNum)
-	wgOut.Add(1)
-	go hti.outputError(errCh)
+	wgOut.Add(2)
+	go hti.outputError(errCh, &wgOut)
 	go hti.outputResult(outCh, &wgOut)
 	for i := 0; i < hti.jobsNum; i++ {
 		go hti.worker(inCh, outCh, errCh, &wg)
@@ -26,6 +26,7 @@ func (hti *HTindex) Run() error {
 	}
 	wg.Wait()
 	close(outCh)
+	close(errCh)
 	wgOut.Wait()
 	return nil
 }
@@ -37,7 +38,6 @@ func (hti *HTindex) readInput(inCh chan<- string, errCh chan<- error) error {
 	}
 	defer file.Close()
 	defer close(inCh)
-	defer close(errCh)
 	scanner := bufio.NewScanner(file)
 	for scanner.Scan() {
 		inCh <- scanner.Text()