+++ worker.go
@@ -4,7 +4,9 @@ import (
 	"archive/zip"
 	"fmt"
 	"io/ioutil"
+	"path/filepath"
 	"regexp"
+	"sort"
 	"strings"
 	"sync"
 
@@ -12,7 +14,14 @@ import (
 	"github.com/gnames/gnfinder/output"
 )
 
-type page struct {
+var isPage = regexp.MustCompile(`\d{8}\.txt`)
+
+type pageContent struct {
+	id   string
+	text []byte
+}
+
+type tpage struct {
 	id         string
 	offset     int
 	offsetNext int
@@ -20,59 +29,72 @@ type page struct {
 
 type title struct {
 	id    string
-	pages []*page
+	pages []*tpage
 	text  []byte
 	res   *output.Output
 }
 
+type byID []*pageContent
+
+func (b byID) Len() int           { return len(b) }
+func (b byID) Less(i, j int) bool { return b[i].id < b[j].id }
+func (b byID) Swap(i, j int)      { b[i], b[j] = b[j], b[i] }
+
 func (hti *HTindex) worker(inCh <-chan string, outCh chan<- *title,
 	errCh chan<- error, wg *sync.WaitGroup) {
-	var pages []*page
-	offset := 0
-	isPage, err := regexp.Compile(`\d{8}\.txt`)
-	if err != nil {
-		errCh <- err
-	}
 	defer wg.Done()
+
 	opts := []gnfinder.Option{
 		gnfinder.OptDict(hti.dict),
 		gnfinder.OptBayes(true),
 	}
 	gnf := gnfinder.NewGNfinder(opts...)
-	_ = gnf
-	for file := range inCh {
-		t := title{id: getID(file), pages: pages}
-		r, err := zip.OpenReader(hti.rootPrefix + file)
+
+	for zipPath := range inCh {
+		offset := 0
+		var pages []*tpage
+		t := title{id: getID(zipPath), pages: pages}
+		r, err := zip.OpenReader(filepath.Join(hti.rootPrefix, zipPath))
 		if err != nil {
 			errCh <- err
 		}
-		for _, f := range r.File {
-			fn := f.Name
-			fnl := len(fn)
-			if fnl < 12 || !isPage.MatchString(fn[fnl-12:fnl]) {
-				continue
-			}
-			// fnEnd := fn[fnl-12 : fnl-4]
-			rc, err := f.Open()
-			if err != nil {
-				errCh <- err
-			}
-			bs, err := ioutil.ReadAll(rc)
-			if err != nil {
-				errCh <- err
-			}
-			p := page{id: fn[fnl-12 : fnl-4], offset: offset}
-			pageUTF := []rune(string(bs))
+
+		for _, pc := range pagesContent(r, errCh) {
+			p := tpage{id: pc.id, offset: offset}
+			pageUTF := []rune(string(pc.text))
 			offset += len(pageUTF)
 			p.offsetNext = offset
 			t.pages = append(t.pages, &p)
-			t.text = append(t.text, bs...)
-			rc.Close()
+			t.text = append(t.text, pc.text...)
 		}
 		r.Close()
 		t.res = gnf.FindNames(t.text)
 		outCh <- &t
 	}
+}
+
+func pagesContent(r *zip.ReadCloser, errCh chan<- error) []*pageContent {
+	var pages []*pageContent
+	for _, f := range r.File {
+		fn := f.Name
+		fnl := len(fn)
+		if fnl < 12 || !isPage.MatchString(fn[fnl-12:fnl]) {
+			continue
+		}
+		zf, err := f.Open()
+		if err != nil {
+			errCh <- err
+		}
+		id := fn[fnl-12 : fnl-4]
+		text, err := ioutil.ReadAll(zf)
+		if err != nil {
+			errCh <- err
+		}
+		pages = append(pages, &pageContent{id: id, text: text})
+		zf.Close()
+	}
+	sort.Sort(byID(pages))
+	return pages
 }
 
 func getID(p string) string {