+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [v0.0.1]
 
+- Add [#5]: documentation in README.md
+- Add [#4]: Save errors and results to filesystem.
 - Add [#3]: Find scientific names in provided content.
 - Add [#2]: Read content data using data from input file.
 - Add [#1]: Read configuration from config file and flags.
@@ -14,6 +16,10 @@ This document follows [changelog guidelines]
 
 [v0.0.1]: https://github.com/gnames/gnfinder/compare/v0.0.0...v0.0.1
 
+[#5]: https://github.com/gnames/htindex/issues/5
+[#4]: https://github.com/gnames/htindex/issues/4
+[#3]: https://github.com/gnames/htindex/issues/3
+[#2]: https://github.com/gnames/htindex/issues/2
 [#1]: https://github.com/gnames/htindex/issues/1
 
 [changelog guidelines]: https://github.com/olivierlacan/keep-a-changelog